% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtils.R
\name{buildEnvPath}
\alias{buildEnvPath}
\title{Build environment path}
\usage{
buildEnvPath(envName, subdir = NULL)
}
\arguments{
\item{envName}{Environment name (string)}

\item{subdir}{Optional subdirectory (string)}
}
\value{
A full file path (string)
}
\description{
Build environment path
}
\keyword{internal}
