% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{VerRapp}
\alias{VerRapp}
\title{Launch the VerR Shiny Application}
\usage{
VerRapp()
}
\value{
A Shiny application object.
}
\description{
This function launches the VerR Shiny application,
 which provides a graphical user interface for managing
 environments and running expressions.
}
\examples{
if (interactive()) {
    VerRapp()
}

}
